[BDN-235] Rename sgmnt package to segment
